+++ crates/amaru-kernel/src/lib.rs
@@ -123,8 +123,8 @@ pub use cardano::{
     lovelace::Lovelace,
     max_string::MaxString128,
     memoized::{
-        self, BorrowedScript, MemoizedDatum, MemoizedNativeScript, MemoizedPlutusData, MemoizedScript,
-        MemoizedTransactionOutput, deserialize_script, serialize_memoized_script, serialize_script,
+        self, BorrowedScript, MemoizedDatum, MemoizedPlutusData, MemoizedScript, MemoizedTransactionOutput,
+        deserialize_script, serialize_memoized_script, serialize_script,
     },
     metadatum::Metadatum,
     multiasset::{self, Multiasset},